+++ inst/example/example.R
@@ -1,4 +1,0 @@
-library(oica)
-
-data(package = 'oica')
-